RndBitmap/RndTex implementation and documentation (#430)
* rndbitmap work

* minor rndtex weak func fix

* bitmap work

* bitmap work

* bitmap work

* slight wav refactor

* more bitmap work

* doc and cleanup

* process flags

* more bitmap matches

* ugh bitmap

* bitmap documentation and cleanup

* tex doc

+++ src/system/utl/WaveFile.h
@@ -21,8 +21,8 @@ public:
     void ReadMarkers();
     void ReadNumSamples();
     IListChunk& PrepareToProvideData();
-    u16 NumChannels() const { return mNumChannels; }
-    u16 BitsPerSample() const { return mBitsPerSample; }
+    int NumChannels() const { return mNumChannels; }
+    int BitsPerSample() const { return mBitsPerSample; }
     int SamplesPerSec() const { return mSamplesPerSec; }
     int NumSamples() const { return mNumSamples; }
     short Format() const { return mFormat; }
